pyproject.toml
[project]
name = "buglens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.6.2",
    "fastapi[all]>=0.128.0",
    "faster-whisper>=0.10.0",
    "ffmpeg-python>=0.2.0",
    "loguru>=0.7.3",
    "opencv-python-headless>=4.12.0.88",
    "pydantic>=2.6.1",
    "python-dotenv>=1.2.1",
    "ultralytics>=8.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
